changed validation.py andvalidation_config.py and constants.py
