+++ dq-framework/Utilities/validation.py
@@ -1,34 +1,11 @@
 # Databricks notebook source
-import json
-import sys
-import os
-from pathlib import Path
-from common.validation_config import REQUIRED_TABLE_METADATA, directory_path, dfs, validation_steps, validations
 from pyspark.sql.functions import *
+from common.validation_config import dfs,metadata,validations
+from common.constants import VALIDATION_STEPS
 
-# Load Metadata
-def load_selected_metadata(directory_path):
-    if not os.path.exists(directory_path):
-        logger.error(f"Directory '{directory_path}' does not exist.")
-        return {}
-    metadata = {}
-    for table_name, filename in REQUIRED_TABLE_METADATA.items():
-        file_path = os.path.join(directory_path, filename)
-        if os.path.exists(file_path):
-            try:
-                with open(file_path, 'r') as file:
-                    table_metadata = json.load(file)
-                    metadata[table_name] = table_metadata
-                    logger.info(f"Loaded metadata for '{table_name}' from '{filename}'")
-            except Exception as e:
-                logger.error(f"Failed to load metadata from '{filename}': {e}")
-        else:
-            logger.error(f"File '{filename}' not found in directory '{directory_path}'")
-    return metadata
-#****************************************************************************************************************
 
-# Fetch Metadata_Property
-
+# Retrieves a specific property from the metadata for a given table.
+# Supports fetching properties such as 'foreign_key', 'primary_key','nullable', and 'type'.
 def fetch_metadata_property(metadata, table_name, property_name):
     try:
         if table_name in metadata:
@@ -55,28 +32,30 @@ def fetch_metadata_property(metadata, table_name, property_name):
             return {}
     except Exception as e:
         logger.info(f"Error fetching '{property_name}' for table '{table_name}': {e}")
-        return {}
-#*********************************************************************************************************************
+        return {}    
 
-# Empty_df validation
 
-def check_empty_dataframe(df,table_name):
+  
+# Validate if a DataFrame is empty for a given table.Logs an error if empty.
+# Ensures that the dataframes contains records before proceeding with further validations.
+def check_empty_dataframe(df,table_name):    
     if df.count() == 0:
         logger.error(f"check_empty_dataframe validation failed: DataFrame for table  {table_name} is empty for entity ID {entity_id}.Please make correct entries in table {table_name} for entity_id {entity_id}")
         return False
     logger.info(f"check_empty_dataframe validation passed: No empty dataframe violations for table{table_name} for entity ID {entity_id}.")
     return True
     
-#*************************************************************************************************************************
 
-# Data_type validation
+
+# Validate data types of each column against the metadata schema.
+# Ensures that the actual data types match the expected types defined in metadata.
 def validate_column_data_types(df, metadata,table_name):
     schema_datatype =  fetch_metadata_property(metadata, table_name,"type")
     if not schema_datatype:
         logger.info(f"validate_column_data_types validation failed: No schema_datatypes exist in table metadata for table {table_name}.Please ensure schema_datatypes exist in the table metadata for table {table_name}.")        
         return False
     df_dtypes = dict(df.dtypes)
-    validation_passed = True
+    validation_passed = True    
     for column, expected_type in schema_datatype.items():
         if column in df_dtypes:
             actual_type = df_dtypes[column]
@@ -90,10 +69,10 @@ def validate_column_data_types(df, metadata,table_name):
         logger.info(f"validate_column_data_types validation passed: All columns in table {table_name} for entity ID {entity_id} have correct data types.")
     return validation_passed
 
-#*****************************************************************************************************************************
 
-# Nullable Validation
 
+# Validate nullable constraints for columns
+# Ensures that columns marked as NOT NULL do not contain NULL values.
 def validate_nullable_constraint(df, metadata,table_name):
     column_constraints = fetch_metadata_property(metadata, table_name,"nullable")
     if not column_constraints:
@@ -110,10 +89,10 @@ def validate_nullable_constraint(df, metadata,table_name):
         logger.info(f"validate_nullable_constraint validation passed: No NULL constraint violations for table {table_name} for entity ID {entity_id}.")
         return validation_passed
 
-#*******************************************************************************************************************************
 
-# Primary key validation
 
+# Validate primary key uniqueness
+# Ensures that primary key columns do not contain duplicate values.
 def validate_primary_key_uniqueness(df, metadata, table_name):
     primary_keys = [key for key, value in fetch_metadata_property(metadata, table_name, "primary_key").items() if value]
     
@@ -140,17 +119,13 @@ def validate_primary_key_uniqueness(df, metadata, table_name):
     return True
 
 
-
-#*********************************************************************************************************************************
-
-# Foreign_key validations
-
+# Validate foreign key relationships
+# Ensures that referenced foreign keys exist in parent tables.
 def validate_foreign_key_relationship(df, metadata, table_name, dfs):
     foreign_keys = fetch_metadata_property(metadata, table_name, "foreign_key")    
     if not foreign_keys:
         logger.info(f"No foreign key relationships defined for table '{table_name}'. Skipping validation")
         return True
-
     for child_column, reference in foreign_keys.items():
         parent_table = reference.get("table")
         parent_column = reference.get("column")
@@ -158,7 +133,7 @@ def validate_foreign_key_relationship(df, metadata, table_name, dfs):
         if parent_table not in dfs:
             logger.error(f"Parent table '{parent_table}' not found in provided DataFrames.")
             return False
-
+           
         parent_df = dfs[parent_table]
 
         missing_rows = df.join(parent_df, df[child_column] == parent_df[parent_column], "left_anti")
@@ -170,19 +145,17 @@ def validate_foreign_key_relationship(df, metadata, table_name, dfs):
                 f"Validation failed: {missing_count} missing foreign key values in '{table_name}.{child_column}' referencing '{parent_table}.{parent_column}'. Missing values: {missing_values}"
             )
             return False
-
     logger.info(f"Validation passed: Foreign key relationships verified for table '{table_name}'.")
     return True
-#************************************************************************************************************************************
 
-# Apply validations
 
+# Apply all validations on each  DataFrame
 def apply_validation(filter_df, metadata, table_name):
     validation_passed = True
     if not check_empty_dataframe(filter_df, table_name):
         logger.error("Validation failed: DataFrame is empty. Skipping further validations.")
         return False
-    for validation_name, validation_func in validation_steps:
+    for validation_name, validation_func in VALIDATION_STEPS:
         try:
             result = validation_func(filter_df, metadata, table_name)
             if not result:
@@ -196,11 +169,10 @@ def apply_validation(filter_df, metadata, table_name):
     else:
         logger.info(f"Validation process failed for table {table_name}.")
     return validation_passed
-#**********************************************************************************************************************************
+    
 
-# Execute Validations
-
-def execute_validations(validations):
+# Execute all validations on given dataframes
+def execute_validations(validations):       
     validation_passed = True
     for validation_func, args in validations:
         if not validation_func(*args):
@@ -210,7 +182,5 @@ def execute_validations(validations):
     if not validation_passed:
         logger.error("Some metadata validations failed. Please check the logs for details.\n Hence Metadata Validation process failed")
         return validation_passed
-    logger.info("Metadata validation process completed successfully.")
+    logger.info("Metadata validation process completed successfully.")       
     return validation_passed
-#********************************************************************************************************************************
-
