+++ dq-framework/common/constants.py
@@ -29,7 +29,23 @@ VAR_S3_EXECUTION_PLAN_PATH = "job_catalog.dq_testdb2.dq_execution_result"
 VAR_S3_EXECUTION_RESULT_PATH = "job_catalog.dq_testdb2.df_rule_master"
 
 # result store paths
-VAR_BAD_RECORD_PATH = "s3://error_record_path/"
-VAR_GOOD_RECORD_PATH = "s3://error_record_path/"
+VAR_BAD_RECORD_PATH = "s3://dq-results-store/bad_records/"
+VAR_GOOD_RECORD_PATH = "s3://dq-results-store/good_records/"
+
+# Directory path containing JSON files
+DIRECTORY_PATH = "dq-framework/metadata"
+
+# Required table metadata
+REQUIRED_TABLE_METADATA = {
+    "dq_entity_master": "dq_entity_master.json",
+    "df_rule_master": "dq_rule_master.json",
+    "dq_execution_plan": "dq_execution_plan.json"
+}
 
 
+# Validation Configuration  required for apply_validation
+VALIDATION_STEPS = [
+    ("Column data type validation", "validate_column_data_types"),
+    ("Nullable constraint validation", "validate_nullable_constraint"),
+    ("Primary key uniqueness validation", "validate_primary_key_uniqueness")
+]
